backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["eathren <nolan.braman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
django-ninja = "^1.3.0"
django-cors-headers = "^4.6.0"
pydantic = "^2.10.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
